# src/modules/coreHttpApi/openapi.yml
openapi: 3.0.0
servers:
  - url: /
    description: Local Connector API
info:
  version: "0.4"
  title: Connector
  description: |
    This API is work in progress and might be changed without further notice on any time.

    You can retrieve the [yaml](/docs/yaml) or [json](/docs/json) representation of this specification by following the corresponding links.

    You can also view this documenation in [rapidoc](/docs/rapidoc).

  license:
    name: MIT
    url: https://raw.githubusercontent.com/nmshd/cns-connector/main/LICENSE
  contact:
    name: j&s-soft GmbH
    email: info@js-soft.com
    url: https://www.js-soft.com/
paths:
...
  /api/v1/Relationships/{id}:
    get:
      operationId: getRelationshipById
      description: Fetches the `Relationship` with the given `id`.
      tags:
        - Relationships
      parameters:
        - in: path
          name: id
          description: The ID of the relationship.
          required: true
          schema:
            $ref: "#/components/schemas/RelationshipID"

      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    nullable: false
                    $ref: "#/components/schemas/Relationship"
                required:
                  - result
          headers:
            X-Response-Duration-ms:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
            X-Response-Time:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Time"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/RelationshipTemplates:
    get:
      operationId: getRelationshipTemplates
      description: Queries `RelationshipTemplates`.
      tags:
        - RelationshipTemplates
      parameters:
        - in: query
          name: createdAt
          schema:
            $ref: "#/components/schemas/DateFilter"
        - in: query
          name: expiresAt
          schema:
            $ref: "#/components/schemas/DateFilter"
        - in: query
          name: createdBy
          schema:
            $ref: "#/components/schemas/IdFilter"
        - in: query
          name: createdByDevice
          schema:
            $ref: "#/components/schemas/IdFilter"
        - in: query
          name: maxNumberOfRelationships
          schema:
            $ref: "#/components/schemas/NumberFilter"
        - in: query
          name: isOwn
          schema:
            $ref: "#/components/schemas/BooleanFilter"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    nullable: false
                    type: array
                    items:
                      $ref: "#/components/schemas/RelationshipTemplate"
                required:
                  - result
          headers:
            X-Response-Duration-ms:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
            X-Response-Time:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Time"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
  /api/v1/RelationshipTemplates/Own:
    post:
      operationId: createOwnRelationshipTemplate
      description: Creates a new `RelationshipTemplate`.
      tags:
        - RelationshipTemplates
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                maxNumberOfRelationships:
                  type: integer
                  format: int32
                  nullable: true
                  description: The maximum number of relationships that can be established with this template. Omitting this property leads to an infinite number of possible relationships.
                  example: 1
                expiresAt:
                  type: string
                  format: date-time
                  example: 2022-01-01
                  description: A timestamp that describes when this relationship template expires. Expired templates cannot be used to create relationship requests anymore.
                content:
                  allOf:
                    - $ref: "#/components/schemas/RelationshipTemplateContent"
                  nullable: false
                  example:
                    prop1: value
                    prop2: 1
              required:
                - expiresAt
                - content
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    nullable: false
                    $ref: "#/components/schemas/RelationshipTemplate"
                required:
                  - result
          headers:
            X-Response-Duration-ms:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
            X-Response-Time:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Time"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"

    get:
      operationId: getOwnRelationshipTemplates
      description: Queries own `RelationshipTemplates`.
      tags:
        - RelationshipTemplates
      parameters:
        - in: query
          name: createdAt
          schema:
            $ref: "#/components/schemas/DateFilter"
        - in: query
          name: expiresAt
          schema:
            $ref: "#/components/schemas/DateFilter"
        - in: query
          name: createdByDevice
          schema:
            $ref: "#/components/schemas/IdFilter"
        - in: query
          name: maxNumberOfRelationships
          schema:
            $ref: "#/components/schemas/NumberFilter"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    nullable: false
                    type: array
                    items:
                      $ref: "#/components/schemas/RelationshipTemplate"
                required:
                  - result
          headers:
            X-Response-Duration-ms:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
            X-Response-Time:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Time"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
  /api/v1/RelationshipTemplates/Peer:
    post:
      operationId: loadPeerRelationshipTemplate
      description: Loads a `RelationshipTemplate` created by others. This is a prerequisite for using the template while creating a new `Relationship`.
      tags:
        - RelationshipTemplates
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/RelationshipTemplateReference"
                - $ref: "#/components/schemas/TokenReferenceTruncated"
            examples:
              RelationshipTemplateReference:
                value: |
                  { 
                    "id": "RLT_________________",
                    "secretKey": "YOUR_SECRET_KEY"
                  }
              TokenReferenceTruncated:
                value: |
                  {
                    "reference": "YOUR_REFERENCE"
                  }
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    nullable: false
                    $ref: "#/components/schemas/RelationshipTemplate"
                required:
                  - result
          headers:
            X-Response-Duration-ms:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
            X-Response-Time:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Time"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"

    get:
      operationId: getPeerRelationshipTemplates
      description: Queries `RelationshipTemplates` of other identities. The templates have to be loaded by calling `POST /RelationshipTemplates/Peer`.
      tags:
        - RelationshipTemplates
      parameters:
        - in: query
          name: createdAt
          schema:
            $ref: "#/components/schemas/DateFilter"
        - in: query
          name: expiresAt
          schema:
            $ref: "#/components/schemas/DateFilter"
        - in: query
          name: createdBy
          schema:
            $ref: "#/components/schemas/IdFilter"
        - in: query
          name: maxNumberOfRelationships
          schema:
            $ref: "#/components/schemas/NumberFilter"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    nullable: false
                    type: array
                    items:
                      $ref: "#/components/schemas/RelationshipTemplate"
                required:
                  - result
          headers:
            X-Response-Duration-ms:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
            X-Response-Time:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Time"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
  /api/v1/RelationshipTemplates/{id}:
    get:
      operationId: getRelationshipTemplate
      description: Fetches the `RelationshipTemplate` with the given `id`.
      tags:
        - RelationshipTemplates
      parameters:
        - in: path
          name: id
          description: The ID of the relationship template.
          required: true
          schema:
            $ref: "#/components/schemas/RelationshipTemplateID"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    nullable: false
                    $ref: "#/components/schemas/RelationshipTemplate"
                required:
                  - result
          headers:
            X-Response-Duration-ms:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
            X-Response-Time:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Time"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
  /api/v1/RelationshipTemplates/Own/{id}/Token:
    post:
      operationId: createTokenForTemplate
      description: Creates a `Token` for the own `RelationshipTemplate` with the given `id`
      tags:
        - RelationshipTemplates
      parameters:
        - in: path
          name: id
          description: The ID of the relationship template.
          required: true
          schema:
            $ref: "#/components/schemas/RelationshipTemplateID"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                expiresAt:
                  description: The expiry of the token. Defaults to the expiry of the relationship template.
                  type: string
                  format: date-time
                ephemeral:
                  description: If set to true the token will will not be cached in the database of the connector. Note that you will not be able to fetch this token unless you remember the id and secretKey of the token. Defaults to true. Will be ignored if Accept is set to image/png.
                  type: boolean
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    nullable: false
                    $ref: "#/components/schemas/Token"
                required:
                  - result
            image/png:
              schema:
                description: A QR Code that contains all necessary information about the token which itself contains a reference to the relationship template.
                type: string
                format: binary
          headers:
            X-Response-Duration-ms:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
            X-Response-Time:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Time"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"

  # ------------------- Tokens -------------------
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
...
    IdFilter:
      type: string
      description: >
        <p>Filters for an ID. Basically a `TextFilter`, but with less useful operators</p>
        <p>The following operators are supported:</p>
        <table>
          <tr>
            <th>Operation</th>
            <th>Example</th>
          </tr>
          <tr>
            <td>equal</td>
            <td>?foo=XXX_________________</td>
          </tr>
          <tr>
            <td>not equal</td>
            <td>?foo=!XXX_________________</td>
          </tr>
          <tr>
            <td>
              exists (=> not undefined)*
              <br/>
            </td>
            <td>?foo=</td>
          </tr>
          <tr>
            <td>not exists (=> undefined)</td>
            <td>?foo=!</td>
          </tr>
          <tr>
            <td>in*</td>
            <td>?foo=XXX_________________&foo=YYY_________________</td>
          </tr>
          <tr>
            <td>not in*</td>
            <td>?foo=!XXX_________________&foo=!YYY_________________</td>
          </tr>
        </table>
        <p>* Due to limitations of Swagger UI, this feature cannot be tested in Swagger UI. Use a REST Client instead (e.g. Postman).</p>
    DateFilter:
      type: string
      description: >
        <p>Filters for a date. In the background, a `DateFilter` compares dates based on their text value. Due to the order of ISO Date string characters though, this doesn't matter - as long as no timezones are used. We are looking for a better solution to this. For now you can convert your non-UTC dates to UTC before sending them in a query.</p>
        <p>The following operators are supported:</p>
        <table>
          <tr>
            <th>Operation</th>
            <th>Example</th>
          </tr>
          <tr>
            <td>equal</td>
            <td>?foo=2020-01-01T00:00:00.000Z</td>
          </tr>
          <tr>
            <td>not equal</td>
            <td>?foo=!2020-01-01T00:00:00.000Z</td>
          </tr>
          <tr>
            <td>
              exists (=> not undefined)*
              <br/>
            </td>
            <td>?foo=</td>
          </tr>
          <tr>
            <td>not exists (=> undefined)</td>
            <td>?foo=!</td>
          </tr>
          <tr>
            <td>greater than</td>
            <td>?foo=>2020</td>
          </tr>
          <tr>
            <td>less than</td>
            <td>?foo=<2020</td>
          </tr>
          <tr>
            <td>greater than or equal to</td>
            <td>?foo=>=2020-01-01T00:00:00.000Z</td>
          </tr>
          <tr>
            <td>less than or equal to</td>
            <td>?foo=<=2020-01-01T00:00:00.000Z</td>
          </tr>
          <tr>
            <td>matches regex</td>
            <td>?foo=~.+01-01T.+</td>
          </tr>
          <tr>
            <td>in*</td>
            <td>?foo=2020-01-01T00:00:00.000Z&foo=2021-01-01T00:00:00.000Z</td>
          </tr>
          <tr>
            <td>not in*</td>
            <td>?foo=!2020-01-01T00:00:00.000Z&foo=!2021-01-01T00:00:00.000Z</td>
          </tr>
        </table>
        <p>* Due to limitations of Swagger UI, this feature cannot be tested in Swagger UI. Use a REST Client instead (e.g. Postman).</p>
    BooleanFilter:
      type: string
      description: >
        <p>Filters for a boolean.
        <p>The following operators are supported:</p>
        <table>
          <tr>
            <th>Operation</th>
            <th>Example</th>
          </tr>
          <tr>
            <td>equal</td>
            <td>?foo=true</td>
          </tr>
          <tr>
            <td>not equal</td>
            <td>?foo=!true</td>
          </tr>
        </table>
        <p>* Due to limitations of Swagger UI, this feature cannot be tested in Swagger UI. Use a REST Client instead (e.g. Postman).</p>
    NumberFilter:
      type: string
      description: >
        <p>Filters for a number.
        <p>The following operators are supported:</p>
        <table>
          <tr>
            <th>Operation</th>
            <th>Example</th>
          </tr>
          <tr>
            <td>equal</td>
            <td>?foo=bar</td>
          </tr>
          <tr>
            <td>not equal</td>
            <td>?foo=!bar</td>
          </tr>
          <tr>
            <td>
              exists (=> not undefined)*
              <br/>
            </td>
            <td>?foo=</td>
          </tr>
          <tr>
            <td>not exists (=> undefined)</td>
            <td>?foo=!</td>
          </tr>
          <tr>
            <td>greater than</td>
            <td>?foo=>5</td>
          </tr>
          <tr>
            <td>less than</td>
            <td>?foo=<5</td>
          </tr>
          <tr>
            <td>greater than or equal to</td>
            <td>?foo=>=5</td>
          </tr>
          <tr>
            <td>less than or equal to</td>
            <td>?foo=5</td>
          </tr>
          <tr>
            <td>in*</td>
            <td>?foo=5&foo=6</td>
          </tr>
          <tr>
            <td>not in*</td>
            <td>?foo=!5&foo=!6</td>
          </tr>
        </table>
        <p>* Due to limitations of Swagger UI, this feature cannot be tested in Swagger UI. Use a REST Client instead (e.g. Postman).</p>

    # ------------------- IDs -------------------
    Address:
      type: string
      format: Address
      minLength: 35
      maxLength: 36
      example: id_________________________________
    DeviceID:
      type: string
      format: DeviceID
      minLength: 20
      maxLength: 20
      example: DVC_________________
...
    RelationshipID:
      type: string
      format: RelationshipID
      minLength: 20
      maxLength: 20
      example: REL_________________
      description: The ID of a relationship.
    RelationshipChangeID:
      type: string
      format: RelationshipChangeID
      minLength: 20
      maxLength: 20
      example: RCH_________________
      description: The ID of a relationship change.
    RelationshipTemplateID:
      type: string
      format: RelationshipTemplateID
      minLength: 20
      maxLength: 20
      example: RLT_________________
      description: The ID of a relationship template.
    TokenID:
      type: string
      format: TokenID
      minLength: 20
      maxLength: 20
      example: TOK_________________
      description: The ID of a message.

    # ------------------- Entities -------------------
...
    Relationship:
      type: object
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/RelationshipID"
          nullable: false
          description: The ID of the relationship.
        template:
          allOf:
            - $ref: "#/components/schemas/RelationshipTemplate"
          nullable: false
        status:
          type: string
          enum: [Pending, Active, Rejected, Revoked, Terminated]
          nullable: false
          description: The status of the relationship
        peer:
          allOf:
            - $ref: "#/components/schemas/Address"
          nullable: false
          description: The address of the peer identity.
        changes:
          type: array
          items:
            $ref: "#/components/schemas/RelationshipChange"
          nullable: false
        lastMessageSentAt:
          type: string
          format: date-time
          nullable: true
          description: A timestamp that describes when the last message was sent to the peer.
        lastMessageReceivedAt:
          type: string
          format: date-time
          nullable: true
          description: A timestamp that describes when the last message was received from the peer.
      required:
        - id
        - template
        - status
        - peer
        - changes
    RelationshipChange:
      type: object
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/RelationshipChangeID"
          nullable: false
        type:
          type: string
          nullable: false
          enum: [Creation, Termination]
        status:
          type: string
          nullable: false
          enum: [Pending, Rejected, Revoked, Accepted]
        request:
          $ref: "#/components/schemas/RelationshipChangeRequest"
        response:
          $ref: "#/components/schemas/RelationshipChangeResponse"
      required:
        - id
        - type
        - status
        - request
    RelationshipChangeRequest:
      type: object
      properties:
        createdBy:
          allOf:
            - $ref: "#/components/schemas/Address"
          nullable: false
        createdByDevice:
          allOf:
            - $ref: "#/components/schemas/DeviceID"
          nullable: false
        createdAt:
          type: string
          format: date-time
          nullable: false
          description: A timestamp that describes when this `RelationshipChange` was created on the platform.
          readOnly: true
        content:
          $ref: "#/components/schemas/RelationshipChangeRequestContent"
      required:
        - createdBy
        - createdByDevice
        - createdAt
        - content
    RelationshipChangeRequestContent:
      type: object
...
    RelationshipTemplateContent:
      type: object
    RelationshipTemplate:
      type: object
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/RelationshipTemplateID"
          nullable: false
          readOnly: true
        maxNumberOfRelationships:
          type: integer
          format: int32
          nullable: true
          description: The maximum number of relationships that can be established with this template. Omitting this property leads to an infinite number of possible relationships.
        isOwn:
          type: boolean
          description: Describes whether the template was created by the connector (true) or whether it was loaded from another identity (false).
          nullable: false
        createdBy:
          allOf:
            - $ref: "#/components/schemas/Address"
          nullable: false
          description: The address of the identity who created this template.
        createdByDevice:
          allOf:
            - $ref: "#/components/schemas/DeviceID"
          nullable: false
          description: The address of the identity who created this template.
        createdAt:
          type: string
          format: date-time
          nullable: false
          description: A timestamp that describes when this relationship template was created on the platform.
        expiresAt:
          type: string
          format: date-time
          description: A timestamp that describes when this relationship template expires. Expired templates cannot be used to create relationship requests anymore.
        content:
          allOf:
            - $ref: "#/components/schemas/RelationshipTemplateContent"
          nullable: false
      required:
        - id
        - isOwn
        - createdBy
        - createdByDevice
        - createdAt
        - expiresAt
        - content
    RelationshipTemplateReference:
      type: object
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/RelationshipTemplateID"
          nullable: false
          description: The ID of the received RelationshipTemplate which should be fetched. This is usually shared within a Token.
        secretKey:
          type: string
          format: byte
          nullable: false
          description: The secret key which was used to encrypt the RelationshipTemplate. This is usually shared within a Token.
      required:
        - id
        - secretKey
    Token:
      type: object
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/TokenID"
          nullable: false
        createdBy:
          allOf:
            - $ref: "#/components/schemas/Address"
          nullable: false
          description: The ID of the identity who created this token.
        createdByDevice:
          allOf:
            - $ref: "#/components/schemas/DeviceID"
          nullable: false
          description: The ID of the device which created this token.
        content:
          allOf:
            - $ref: "#/components/schemas/TokenContent"
          nullable: false
        createdAt:
          type: string
          format: date-time
          nullable: false
          description: A timestamp that describes when this token was created on the platform.
        expiresAt:
          type: string
          format: date-time
          nullable: false
          description: A timestamp that describes when this token expires. An expired token cannot be fetched from the platform anymore. However it will still be available for auditing purposes.
        secretKey:
          type: string
          format: byte
          nullable: false
          description: The base64 encoded secret key which was used to encrypt the Token. This is usually shared over the side channel (QR-Code, Link).
        truncatedReference:
          type: string
          format: byte
          nullable: false
          description: The base64 encoded truncated reference of the token, which actually consists of the TokenId and the secretKey.
      required:
        - id
        - createdBy
        - createdByDevice
        - content
        - createdAt
        - expiresAt
        - secretKey
        - truncatedReference
...
    TokenReferenceTruncated:
      type: object
      properties:
        reference:
          type: string
          format: byte
          nullable: false
          description: The base64 encoded truncated reference of the token, which actually consists of the TokenId and the secretKey.
      required:
        - reference
    TokenContent:
      type: object
      description: The arbitrary JSON object which should be shared between creator of the Token and the recipient.

    # ------------------- General -------------------
    ErrorContent:
      type: object
      properties:
        error:
          type: object
          properties:
            id:
              description: A unique ID for the error instance.
              type: string
            code:
              description: >
                The error code.  

                There are some generic error codes like: 
                  - error.connector.recordNotFound
                  - error.connector.unauthorized
                  - error.connector.forbidden
                  - error.connector.validation.missingProperty
                  - error.connector.validation.invalidPropertyValueLength
                  - error.connector.validation.invalidPropertyValue

                Further there are route specific error codes.
              type: string
            message:
              description: An error message having more details about what went wrong. This message is not localized.
              type: string
            docs:
              description: A link to additional documentation about the error.
              type: string
            time:
              description: A timestamp that describes when the error occured.
              type: string
          required:
            - id
            - code
            - message
            - docs
            - time
      required:
        - error
    HeaderContent_X-Response-Duration-ms:
      description: The time in milliseconds the request took to be processed by the platform.
      type: number
    HeaderContent_X-Response-Time:
      description: A timestamp that describes when the response was sent to the client.
      type: string
      format: number
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorContent"
      headers:
        X-Response-Duration-ms:
          schema:
            $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
        X-Response-Time:
          schema:
            $ref: "#/components/schemas/HeaderContent_X-Response-Time"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorContent"
      headers:
        X-Response-Duration-ms:
          schema:
            $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
        X-Response-Time:
          schema:
            $ref: "#/components/schemas/HeaderContent_X-Response-Time"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorContent"
      headers:
        X-Response-Duration-ms:
          schema:
            $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
        X-Response-Time:
          schema:
            $ref: "#/components/schemas/HeaderContent_X-Response-Time"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorContent"
      headers:
        X-Response-Duration-ms:
          schema:
            $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
        X-Response-Time:
          schema:
            $ref: "#/components/schemas/HeaderContent_X-Response-Time"